Simplify Trait classes.
