+++ Assets/Scripts/Characters/CharacterBasicData.cs
@@ -13,6 +13,6 @@ namespace GuildMaster.Characters
         public Sprite Illustration;
         public List<string> NameList; //이름의 리스트 (가장 마지막 것이 진짜이름)
         public string RealName => NameList.ElementAtOrDefault(NameList.Count - 1);
-        public List<CharacterTraitData.Trait> ActiveTraits;
+        public List<Trait> ActiveTraits;
     }
 }
