--- Assets/Scripts/Characters/CharacterTraitData.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: ce6f1cf9f5f91eb4a9b1c6685200b0ea
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
