+++ Assets/Scripts/Characters/Trait.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: b578386c21874b298e0c1c94cedb9c39
+timeCreated: 1603798507
