+++ Assets/Scripts/Databases/TraitDatabase.cs
@@ -1,10 +1,11 @@
+using GuildMaster.Characters;
 using UnityEditor;
 using UnityEngine;
 
 namespace GuildMaster.Databases
 {
     [CreateAssetMenu(fileName = "TraitDatabase", menuName = "ScriptableObjects/TraitDatabase", order = 0)]
-    public class TraitDatabase: UnityEditableEnumDatabase<TraitDatabase, CharacterTraitData.TRAITDATA, CharacterTraitData.Trait>
+    public class TraitDatabase: UnityEditableEnumDatabase<TraitDatabase, TraitStaticData, Trait>
     {}
 
     [CustomEditor(typeof(TraitDatabase))]
